+++ src/pipeline/mod.rs
@@ -1,7 +1,7 @@
 //! Persistent and time-coherent collision detection.
 
 pub mod broad_phase;
-pub mod events;
 pub mod narrow_phase;
 pub mod world;
-pub mod world2;
+pub mod glue;
+pub mod object;
