+++ src/pipeline/narrow_phase/interaction_graph.rs
@@ -3,7 +3,6 @@ use petgraph::visit::EdgeRef;
 use na::RealField;
 
 use crate::query::{ContactManifold, Proximity};
-use crate::world::CollisionObjectHandle;
 use crate::pipeline::narrow_phase::{ContactAlgorithm, ProximityAlgorithm};
 use petgraph::prelude::EdgeIndex;
 use petgraph::Direction;
@@ -46,12 +45,37 @@ impl<N: RealField> Interaction<N> {
 }
 
 /// A graph where nodes are collision objects and edges are contact or proximity algorithms.
-pub struct InteractionGraph<N: RealField>(pub(crate) UnGraph<CollisionObjectHandle, Interaction<N>, usize>);
+pub struct InteractionGraph<N: RealField, Handle: Copy>(pub(crate) UnGraph<Handle, Interaction<N>, usize>);
 
-impl<N: RealField> InteractionGraph<N> {
+impl<N: RealField, Handle: Copy> InteractionGraph<N, Handle> {
     /// Creates a new empty collection of collision objects.
     pub fn new() -> Self {
         InteractionGraph(UnGraph::with_capacity(10, 10))
+    }
+
+    /// Adds a handle to this graph.
+    pub fn add_node(&mut self, handle: Handle) -> CollisionObjectGraphIndex {
+        self.0.add_node(handle)
+    }
+
+    /// Removes a handle from this graph and returns a handle that must have its graph index changed to `id`.
+    ///
+    /// When a node is removed, another node of the graph takes it place. This means that the `CollisionObjectGraphIndex`
+    /// of the collision object returned by this method will be equal to `id`. Thus if you maintain
+    /// a map between `CollisionObjectHandle` and `CollisionObjectGraphIndex`, then you should update this
+    /// map to associate `id` to the handle returned by this method. For example:
+    ///
+    /// ```
+    /// // Let `id` be the graph index of the collision object we want to remove.
+    /// if let Some(other_handle) = graph.remove_node(id) {
+    ///    // The graph index of `other_handle` changed to `id` due to the removal.
+    ///    map.insert(other_handle, id) ;
+    /// }
+    /// ```
+    #[must_use = "The graph index of the collision object returned by this method has been changed to `id`."]
+    pub fn remove_node(&mut self, id: CollisionObjectGraphIndex) -> Option<Handle> {
+        let _ = self.0.remove_node(id);
+        self.0.node_weight(id).cloned()
     }
 
     /// All the interactions pairs on this graph.
@@ -59,8 +83,8 @@ impl<N: RealField> InteractionGraph<N> {
     /// Refer to the official [user guide](https://ncollide.org/interaction_handling_and_sensors/#interaction-iterators)
     /// for details.
     pub fn interaction_pairs(&self, effective_only: bool) -> impl Iterator<Item = (
-        CollisionObjectHandle,
-        CollisionObjectHandle,
+        Handle,
+        Handle,
         &Interaction<N>
     )> {
         self.0
@@ -81,8 +105,8 @@ impl<N: RealField> InteractionGraph<N> {
     /// Refer to the official [user guide](https://ncollide.org/interaction_handling_and_sensors/#interaction-iterators)
     /// for details.
     pub fn contact_pairs(&self, effective_only: bool) -> impl Iterator<Item = (
-        CollisionObjectHandle,
-        CollisionObjectHandle,
+        Handle,
+        Handle,
         &ContactAlgorithm<N>,
         &ContactManifold<N>,
     )> {
@@ -100,8 +124,8 @@ impl<N: RealField> InteractionGraph<N> {
     /// Refer to the official [user guide](https://ncollide.org/interaction_handling_and_sensors/#interaction-iterators)
     /// for details.
     pub fn proximity_pairs(&self, effective_only: bool) -> impl Iterator<Item = (
-        CollisionObjectHandle,
-        CollisionObjectHandle,
+        Handle,
+        Handle,
         &ProximityAlgorithm<N>,
     )> {
         self.interaction_pairs(effective_only)
@@ -117,7 +141,7 @@ impl<N: RealField> InteractionGraph<N> {
     ///
     /// Refer to the official [user guide](https://ncollide.org/interaction_handling_and_sensors/#interaction-iterators)
     /// for details.
-    pub fn interaction_pair(&self, id1: CollisionObjectGraphIndex, id2: CollisionObjectGraphIndex, effective_only: bool) -> Option<(CollisionObjectHandle, CollisionObjectHandle, &Interaction<N>)> {
+    pub fn interaction_pair(&self, id1: CollisionObjectGraphIndex, id2: CollisionObjectGraphIndex, effective_only: bool) -> Option<(Handle, Handle, &Interaction<N>)> {
         let inter = self.0.find_edge(id1, id2).and_then(|edge| {
             let endpoints = self.0.edge_endpoints(edge)?;
             let h1 = self.0.node_weight(endpoints.0)?;
@@ -137,7 +161,7 @@ impl<N: RealField> InteractionGraph<N> {
     ///
     /// Refer to the official [user guide](https://ncollide.org/interaction_handling_and_sensors/#interaction-iterators)
     /// for details.
-    pub fn contact_pair(&self, id1: CollisionObjectGraphIndex, id2: CollisionObjectGraphIndex, effective_only: bool) -> Option<(CollisionObjectHandle, CollisionObjectHandle, &ContactAlgorithm<N>, &ContactManifold<N>)> {
+    pub fn contact_pair(&self, id1: CollisionObjectGraphIndex, id2: CollisionObjectGraphIndex, effective_only: bool) -> Option<(Handle, Handle, &ContactAlgorithm<N>, &ContactManifold<N>)> {
         self.interaction_pair(id1, id2, effective_only).and_then(|inter| {
             match inter.2 {
                 Interaction::Contact(algo, manifold) => Some((inter.0, inter.1, algo, manifold)),
@@ -150,7 +174,7 @@ impl<N: RealField> InteractionGraph<N> {
     ///
     /// Refer to the official [user guide](https://ncollide.org/interaction_handling_and_sensors/#interaction-iterators)
     /// for details.
-    pub fn proximity_pair(&self, id1: CollisionObjectGraphIndex, id2: CollisionObjectGraphIndex, effective_only: bool) -> Option<(CollisionObjectHandle, CollisionObjectHandle, &ProximityAlgorithm<N>)> {
+    pub fn proximity_pair(&self, id1: CollisionObjectGraphIndex, id2: CollisionObjectGraphIndex, effective_only: bool) -> Option<(Handle, Handle, &ProximityAlgorithm<N>)> {
         self.interaction_pair(id1, id2, effective_only).and_then(|inter| {
             match inter.2 {
                 Interaction::Proximity(algo) => Some((inter.0, inter.1, algo)),
@@ -163,7 +187,7 @@ impl<N: RealField> InteractionGraph<N> {
     ///
     /// Refer to the official [user guide](https://ncollide.org/interaction_handling_and_sensors/#interaction-iterators)
     /// for details.
-    pub fn interactions_with(&self, id: CollisionObjectGraphIndex, effective_only: bool) -> impl Iterator<Item = (CollisionObjectHandle, CollisionObjectHandle, &Interaction<N>)> {
+    pub fn interactions_with(&self, id: CollisionObjectGraphIndex, effective_only: bool) -> impl Iterator<Item = (Handle, Handle, &Interaction<N>)> {
         self.0.edges(id).filter_map(move |e| {
             let inter = e.weight();
 
@@ -177,7 +201,7 @@ impl<N: RealField> InteractionGraph<N> {
     }
 
     /// Gets the interaction with the given index.
-    pub fn index_interaction(&self, id: TemporaryInteractionIndex) -> Option<(CollisionObjectHandle, CollisionObjectHandle, &Interaction<N>)> {
+    pub fn index_interaction(&self, id: TemporaryInteractionIndex) -> Option<(Handle, Handle, &Interaction<N>)> {
         if let (Some(e), Some(endpoints)) = (self.0.edge_weight(id), self.0.edge_endpoints(id)) {
             Some((self.0[endpoints.0], self.0[endpoints.1], e))
         } else {
@@ -189,7 +213,7 @@ impl<N: RealField> InteractionGraph<N> {
     ///
     /// Refer to the official [user guide](https://ncollide.org/interaction_handling_and_sensors/#interaction-iterators)
     /// for details.
-    pub fn interactions_with_mut(&mut self, id: CollisionObjectGraphIndex) -> impl Iterator<Item = (CollisionObjectHandle, CollisionObjectHandle, TemporaryInteractionIndex, &mut Interaction<N>)> {
+    pub fn interactions_with_mut(&mut self, id: CollisionObjectGraphIndex) -> impl Iterator<Item = (Handle, Handle, TemporaryInteractionIndex, &mut Interaction<N>)> {
         let incoming_edge = self.0.first_edge(id, Direction::Incoming);
         let outgoing_edge = self.0.first_edge(id, Direction::Outgoing);
 
@@ -204,7 +228,7 @@ impl<N: RealField> InteractionGraph<N> {
     ///
     /// Refer to the official [user guide](https://ncollide.org/interaction_handling_and_sensors/#interaction-iterators)
     /// for details.
-    pub fn proximities_with(&self, handle: CollisionObjectGraphIndex, effective_only: bool) -> impl Iterator<Item = (CollisionObjectHandle, CollisionObjectHandle, &ProximityAlgorithm<N>)> {
+    pub fn proximities_with(&self, handle: CollisionObjectGraphIndex, effective_only: bool) -> impl Iterator<Item = (Handle, Handle, &ProximityAlgorithm<N>)> {
         self.interactions_with(handle, effective_only)
             .filter_map(|(h1, h2, inter)| {
                 match inter {
@@ -219,7 +243,7 @@ impl<N: RealField> InteractionGraph<N> {
     ///
     /// Refer to the official [user guide](https://ncollide.org/interaction_handling_and_sensors/#interaction-iterators)
     /// for details.
-    pub fn contacts_with(&self, handle: CollisionObjectGraphIndex, effective_only: bool) -> impl Iterator<Item = (CollisionObjectHandle, CollisionObjectHandle, &ContactAlgorithm<N>, &ContactManifold<N>)> {
+    pub fn contacts_with(&self, handle: CollisionObjectGraphIndex, effective_only: bool) -> impl Iterator<Item = (Handle, Handle, &ContactAlgorithm<N>, &ContactManifold<N>)> {
         self.interactions_with(handle, effective_only)
             .filter_map(|(h1, h2, inter)| {
                 match inter {
@@ -234,7 +258,7 @@ impl<N: RealField> InteractionGraph<N> {
     ///
     /// Refer to the official [user guide](https://ncollide.org/interaction_handling_and_sensors/#interaction-iterators)
     /// for details.
-    pub fn collision_objects_interacting_with<'a>(&'a self, id: CollisionObjectGraphIndex) -> impl Iterator<Item = CollisionObjectHandle> + 'a {
+    pub fn collision_objects_interacting_with<'a>(&'a self, id: CollisionObjectGraphIndex) -> impl Iterator<Item = Handle> + 'a {
         self.0.edges(id).filter_map(move |e| {
             let inter = e.weight();
 
@@ -254,7 +278,7 @@ impl<N: RealField> InteractionGraph<N> {
     ///
     /// Refer to the official [user guide](https://ncollide.org/interaction_handling_and_sensors/#interaction-iterators)
     /// for details.
-    pub fn collision_objects_in_contact_with<'a>(&'a self, id: CollisionObjectGraphIndex) -> impl Iterator<Item = CollisionObjectHandle> + 'a {
+    pub fn collision_objects_in_contact_with<'a>(&'a self, id: CollisionObjectGraphIndex) -> impl Iterator<Item = Handle> + 'a {
         self.0.edges(id).filter_map(move |e| {
             let inter = e.weight();
 
@@ -274,7 +298,7 @@ impl<N: RealField> InteractionGraph<N> {
     ///
     /// Refer to the official [user guide](https://ncollide.org/interaction_handling_and_sensors/#interaction-iterators)
     /// for details.
-    pub fn collision_objects_in_proximity_of<'a>(&'a self, id: CollisionObjectGraphIndex) -> impl Iterator<Item = CollisionObjectHandle> + 'a {
+    pub fn collision_objects_in_proximity_of<'a>(&'a self, id: CollisionObjectGraphIndex) -> impl Iterator<Item = Handle> + 'a {
         self.0.edges(id).filter_map(move |e| {
             if let Interaction::Proximity(alg) = e.weight() {
                 if alg.proximity() == Proximity::Intersecting {
@@ -308,17 +332,17 @@ impl<N: RealField> InteractionGraph<N> {
 }
 
 
-pub struct InteractionsWithMut<'a, N: RealField> {
-    graph: &'a mut InteractionGraph<N>,
+pub struct InteractionsWithMut<'a, N: RealField, Handle: Copy> {
+    graph: &'a mut InteractionGraph<N, Handle>,
     incoming_edge: Option<EdgeIndex<usize>>,
     outgoing_edge: Option<EdgeIndex<usize>>,
 }
 
-impl<'a, N: RealField> Iterator for InteractionsWithMut<'a, N> {
-    type Item = (CollisionObjectHandle, CollisionObjectHandle, TemporaryInteractionIndex, &'a mut Interaction<N>);
+impl<'a, N: RealField, Handle: Copy> Iterator for InteractionsWithMut<'a, N, Handle> {
+    type Item = (Handle, Handle, TemporaryInteractionIndex, &'a mut Interaction<N>);
 
     #[inline]
-    fn next(&mut self) -> Option<(CollisionObjectHandle, CollisionObjectHandle, TemporaryInteractionIndex, &'a mut Interaction<N>)> {
+    fn next(&mut self) -> Option<(Handle, Handle, TemporaryInteractionIndex, &'a mut Interaction<N>)> {
         if let Some(edge) = self.incoming_edge {
             self.incoming_edge = self.graph.0.next_edge(edge, Direction::Incoming);
             let endpoints = self.graph.0.edge_endpoints(edge).unwrap();
