+++ src/pipeline/narrow_phase/narrow_phase.rs
@@ -1,31 +1,30 @@
 use na::RealField;
 use slotmap::{Key, SlotMap};
 
-use crate::pipeline::events::{ContactEvent, ContactEvents, ProximityEvent, ProximityEvents};
 use crate::pipeline::narrow_phase::{
+    ContactEvent, ContactEvents, ProximityEvent, ProximityEvents,
     ContactDispatcher, ProximityDispatcher, InteractionGraph, Interaction, CollisionObjectGraphIndex,
     ContactManifoldGenerator, ProximityDetector,
 };
-use crate::pipeline::world::{CollisionObjectHandle, CollisionObjectSlab, CollisionObject, GeometricQueryType};
 use crate::query::{Proximity, ContactManifold, ContactId};
-use crate::utils::SortedPair;
+use crate::pipeline::object::{CollisionObjectRef, CollisionObjectSet, GeometricQueryType};
 
-// FIXME: move this to the `narrow_phase` module.
+
 /// Collision detector dispatcher for collision objects.
-pub struct NarrowPhase<N: RealField> {
+pub struct NarrowPhase<N: RealField, Handle: Copy> {
     contact_dispatcher: Box<ContactDispatcher<N>>,
     proximity_dispatcher: Box<ProximityDispatcher<N>>,
-    contact_events: ContactEvents,
-    proximity_events: ProximityEvents,
+    contact_events: ContactEvents<Handle>,
+    proximity_events: ProximityEvents<Handle>,
     id_allocator: SlotMap<ContactId, bool>,
 }
 
-impl<N: RealField> NarrowPhase<N> {
+impl<N: RealField, Handle: Copy> NarrowPhase<N, Handle> {
     /// Creates a new `NarrowPhase`.
     pub fn new(
         contact_dispatcher: Box<ContactDispatcher<N>>,
         proximity_dispatcher: Box<ProximityDispatcher<N>>,
-    ) -> NarrowPhase<N>
+    ) -> NarrowPhase<N, Handle>
     {
         NarrowPhase {
             contact_dispatcher,
@@ -36,7 +35,7 @@ impl<N: RealField> NarrowPhase<N> {
         }
     }
 
-    fn garbage_collect_ids(&mut self, interactions: &mut InteractionGraph<N>) {
+    fn garbage_collect_ids(&mut self, interactions: &mut InteractionGraph<N, Handle>) {
         for interaction in interactions.0.edge_weights_mut() {
             match interaction {
                 Interaction::Contact(_, manifold) => {
@@ -55,12 +54,11 @@ impl<N: RealField> NarrowPhase<N> {
         })
     }
 
-
     /// Update the specified contact manifold between two collision objects.
-    pub fn update_contact<T>(
+    pub fn update_contact<'a>(
         &mut self,
-        co1: &CollisionObject<N, T>,
-        co2: &CollisionObject<N, T>,
+        co1: impl CollisionObjectRef<'a, N, Handle = Handle>,
+        co2: impl CollisionObjectRef<'a, N, Handle = Handle>,
         detector: &mut ContactManifoldGenerator<N>,
         manifold: &mut ContactManifold<N>) {
         let had_contacts = manifold.len() != 0;
@@ -73,10 +71,10 @@ impl<N: RealField> NarrowPhase<N> {
             let _ = detector.generate_contacts(
                 &*self.contact_dispatcher,
                 &co1.position(),
-                co1.shape().as_ref(),
+                co1.shape(),
                 None,
                 &co2.position(),
-                co2.shape().as_ref(),
+                co2.shape(),
                 None,
                 &prediction,
                 manifold,
@@ -103,19 +101,19 @@ impl<N: RealField> NarrowPhase<N> {
     }
 
     /// Update the specified proximity between two collision objects.
-    pub fn update_proximity<T>(
+    pub fn update_proximity<'a>(
         &mut self,
-        co1: &CollisionObject<N, T>,
-        co2: &CollisionObject<N, T>,
+        co1: impl CollisionObjectRef<'a, N, Handle = Handle>,
+        co2: impl CollisionObjectRef<'a, N, Handle = Handle>,
         detector: &mut ProximityDetector<N>) {
         let prev_prox = detector.proximity();
 
         let _ = detector.update(
             &*self.proximity_dispatcher,
             &co1.position(),
-            co1.shape().as_ref(),
+            co1.shape(),
             &co2.position(),
-            co2.shape().as_ref(),
+            co2.shape(),
             co1.query_type().query_limit() + co2.query_type().query_limit(),
         );
 
@@ -132,10 +130,10 @@ impl<N: RealField> NarrowPhase<N> {
     }
 
     /// Update the specified interaction between two collision objects.
-    pub fn update_interaction<T>(
+    pub fn update_interaction<'a>(
         &mut self,
-        co1: &CollisionObject<N, T>,
-        co2: &CollisionObject<N, T>,
+        co1: impl CollisionObjectRef<'a, N, Handle = Handle>,
+        co2: impl CollisionObjectRef<'a, N, Handle = Handle>,
         interaction: &mut Interaction<N>) {
         match interaction {
             Interaction::Contact(detector, manifold) => {
@@ -151,36 +149,48 @@ impl<N: RealField> NarrowPhase<N> {
     /// interactions pairs reported by the broad-phase.
     ///
     /// This will push relevant events to `contact_events` and `proximity_events`.
-    pub fn update<T>(&mut self, interactions: &mut InteractionGraph<N>, objects: &CollisionObjectSlab<N, T>, timestamp: usize, )
+    pub fn update<'a, Objects>(
+        &mut self,
+        interactions: &mut InteractionGraph<N, Objects::Handle>,
+        objects: &'a Objects)
+        where Objects: CollisionObjectSet<'a, N, Handle = Handle>
     {
         for eid in interactions.0.edge_indices() {
             let (id1, id2) = interactions.0.edge_endpoints(eid).unwrap();
-            let co1 = &objects[interactions.0[id1]];
-            let co2 = &objects[interactions.0[id2]];
+            let co1 = objects.get(interactions.0[id1]).unwrap();
+            let co2 = objects.get(interactions.0[id2]).unwrap();
+            let flags1 = co1.update_flags();
+            let flags2 = co2.update_flags();
 
-            if co1.timestamp == timestamp || co2.timestamp == timestamp {
+            if flags1.needs_narrow_phase_update() || flags2.needs_narrow_phase_update() {
                 self.update_interaction(co1, co2, interactions.0.edge_weight_mut(eid).unwrap())
             }
         }
 
-        self.garbage_collect_ids(interactions);
+        // FIXME: don't do this at each update?
+        self.garbage_collect_ids(interactions)
     }
 
     /// Handles a pair of collision objects detected as either started or stopped interacting.
-    pub fn handle_interaction<T>(
+    pub fn handle_interaction<'a, Objects>(
         &mut self,
-        interactions: &mut InteractionGraph<N>,
-        objects: &CollisionObjectSlab<N, T>,
-        handle1: CollisionObjectHandle,
-        handle2: CollisionObjectHandle,
+        interactions: &mut InteractionGraph<N, Objects::Handle>,
+        objects: &'a Objects,
+        handle1: Objects::Handle,
+        handle2: Objects::Handle,
         started: bool,
     )
+        where Objects: CollisionObjectSet<'a, N, Handle = Handle>
     {
-        let key = SortedPair::new(handle1, handle2);
-        let co1 = &objects[key.0];
-        let co2 = &objects[key.1];
-        let id1 = co1.graph_index();
-        let id2 = co2.graph_index();
+        let mut co1 = objects.get(handle1).unwrap();
+        let mut co2 = objects.get(handle2).unwrap();
+        let mut id1 = co1.graph_index();
+        let mut id2 = co2.graph_index();
+
+        if id1 > id2 {
+            std::mem::swap(&mut co1, &mut co2);
+            std::mem::swap(&mut id1, &mut id2);
+        }
 
         if started {
             if !interactions.0.contains_edge(id1, id2) {
@@ -189,7 +199,7 @@ impl<N: RealField> NarrowPhase<N> {
                         let dispatcher = &self.contact_dispatcher;
 
                         if let Some(detector) = dispatcher
-                            .get_contact_algorithm(co1.shape().as_ref(), co2.shape().as_ref())
+                            .get_contact_algorithm(co1.shape(), co2.shape())
                             {
                                 let manifold = detector.init_manifold();
                                 let _ = interactions.0.add_edge(id1, id2, Interaction::Contact(detector, manifold));
@@ -199,7 +209,7 @@ impl<N: RealField> NarrowPhase<N> {
                         let dispatcher = &self.proximity_dispatcher;
 
                         if let Some(detector) = dispatcher
-                            .get_proximity_algorithm(co1.shape().as_ref(), co2.shape().as_ref())
+                            .get_proximity_algorithm(co1.shape(), co2.shape())
                             {
                                 let _ = interactions.0.add_edge(id1, id2, Interaction::Proximity(detector));
                             }
@@ -238,42 +248,13 @@ impl<N: RealField> NarrowPhase<N> {
         }
     }
 
-    /// Handles the addition of a new collision object.
-    pub fn handle_collision_object_added(
-        &mut self,
-        interactions: &mut InteractionGraph<N>,
-        object: CollisionObjectHandle
-    ) -> CollisionObjectGraphIndex {
-        interactions.0.add_node(object)
-    }
-
-    /// Handles the removal of a collision object.
-    pub fn handle_collision_object_removed<T>(
-        &mut self,
-        interactions: &mut InteractionGraph<N>,
-        object: &CollisionObject<N, T>
-    ) -> Option<CollisionObjectHandle> {
-        let id = object.graph_index();
-        let mut nbhs = interactions.0.neighbors(id).detach();
-
-        // Clear all the manifold to avoid leaking contact IDs.
-        while let Some((eid, _)) = nbhs.next(&interactions.0) {
-            match interactions.0.edge_weight_mut(eid).unwrap() {
-                Interaction::Contact(_, manifold) => manifold.clear(),
-                Interaction::Proximity(_) => {}
-            }
-        }
-
-        interactions.0.remove_node(object.graph_index())
-    }
-
     /// The set of contact events generated by this narrow-phase.
-    pub fn contact_events(&self) -> &ContactEvents {
+    pub fn contact_events(&self) -> &ContactEvents<Handle> {
         &self.contact_events
     }
 
     /// The set of proximity events generated by this narrow-phase.
-    pub fn proximity_events(&self) -> &ProximityEvents {
+    pub fn proximity_events(&self) -> &ProximityEvents<Handle> {
         &self.proximity_events
     }
