+++ src/pipeline/object/collision_groups.rs
@@ -1,8 +1,6 @@
 use na::RealField;
 use crate::pipeline::broad_phase::BroadPhasePairFilter;
-use crate::pipeline::broad_phase::BroadPhasePairFilter2;
-use crate::pipeline::world::CollisionObject;
-use crate::pipeline::world2::CollisionObjectRef;
+use crate::pipeline::object::CollisionObjectRef;
 
 const SELF_COLLISION: u32 = 1 << 31;
 const ALL_GROUPS: u32 = (1 << 30) - 1;
@@ -264,6 +262,7 @@ impl Default for CollisionGroups {
     }
 }
 
+
 /// A collision filter based collision groups.
 pub struct CollisionGroupsPairFilter;
 
@@ -275,31 +274,7 @@ impl CollisionGroupsPairFilter {
     }
 }
 
-impl<N: RealField, T> BroadPhasePairFilter<N, T> for CollisionGroupsPairFilter {
-    fn is_pair_valid(&self, co1: &CollisionObject<N, T>, co2: &CollisionObject<N, T>) -> bool {
-        if co1.handle() == co2.handle() {
-            co1.collision_groups().can_interact_with_self()
-        } else {
-            co1.collision_groups()
-                .can_interact_with_groups(co2.collision_groups())
-        }
-    }
-}
-
-
-
-/// A collision filter based collision groups.
-pub struct CollisionGroupsPairFilter2;
-
-impl CollisionGroupsPairFilter2 {
-    /// Creates a new collision filter based collision groups.
-    #[inline]
-    pub fn new() -> CollisionGroupsPairFilter2 {
-        CollisionGroupsPairFilter2
-    }
-}
-
-impl<'a, N: RealField, Object: CollisionObjectRef<'a, N>> BroadPhasePairFilter2<'a, N, Object> for CollisionGroupsPairFilter2 {
+impl<'a, N: RealField, Object: CollisionObjectRef<'a, N>> BroadPhasePairFilter<N, Object> for CollisionGroupsPairFilter {
     fn is_pair_valid(&self, co1: Object, co2: Object) -> bool {
         if co1.is_same_as(co2) {
             co1.collision_groups().can_interact_with_self()
