+++ src/pipeline/object/mod.rs
@@ -1,0 +1,12 @@
+pub use self::collision_groups::{CollisionGroups, CollisionGroupsPairFilter};
+pub use self::collision_object::{
+    CollisionObjectRef, CollisionObject, CollisionObjectHandle,
+};
+pub use self::collision_object_set::{CollisionObjectSet, CollisionObjectSlab, CollisionObjects};
+pub use self::query_type::GeometricQueryType;
+
+
+mod collision_object;
+mod collision_object_set;
+mod collision_groups;
+mod query_type;
